requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_py-xgboost-mutex=2.0=cpu_0
appnope=0.1.2=py39hecd8cb5_1001
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
boost=1.74.0=py39ha641261_3
boost-cpp=1.74.0=h43a636a_2
bzip2=1.0.8=h0d85af4_4
ca-certificates=2021.5.25=hecd8cb5_1
cairo=1.16.0=he43a7df_1008
certifi=2020.12.5=py39h6e9494a_1
click=7.1.2=pypi_0
cycler=0.10.0=py39hecd8cb5_0
decorator=5.0.6=pyhd3eb1b0_0
et_xmlfile=1.0.1=py_1001
fire=0.4.0=pyh44b312d_0
fontconfig=2.13.1=h10f422b_1005
freetype=2.10.4=ha233b18_0
gettext=0.21.0=h7535e17_0
greenlet=1.1.0=py39h9fcab8e_0
icu=68.1=h74dc148_0
intel-openmp=2021.2.0=hecd8cb5_564
ipykernel=5.3.4=py39h01d92e1_0
ipython=7.22.0=py39h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jdcal=1.4.1=py_0
jedi=0.17.2=py39hecd8cb5_1
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=he5867d9_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py39hecd8cb5_0
kallisto=1.0.7=pypi_0
kiwisolver=1.3.1=py39h23ab428_0
lcms2=2.12=hf1fd2bf_0
libcxx=11.1.0=habf9029_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libglib=2.68.1=hd556434_0
libiconv=1.16=haf1e3a3_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libxgboost=1.3.3=h23ab428_0
libxml2=2.9.10=h93ec3fd_4
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.3=h23ab428_0
matplotlib=3.3.4=py39hecd8cb5_0
matplotlib-base=3.3.4=py39h8b3ea08_0
mkl=2021.2.0=hecd8cb5_269
mkl-service=2.3.0=py39h9ed2024_1
mkl_fft=1.3.0=py39h4a7008c_2
mkl_random=1.2.1=py39hb2f4e1b_2
morfeus-ml=0.5.3=pyhd8ed1ab_0
ncurses=6.2=h0a44026_1
numpy=1.20.1=py39hd6e1bb9_0
numpy-base=1.20.1=py39h585ceec_0
olefile=0.46=py_0
openpyxl=3.0.7=pyhd3eb1b0_0
openssl=1.1.1k=h0d85af4_0
pandas=1.2.4=py39h23ab428_0
parso=0.7.0=py_0
pcre=8.44=hb1e8313_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py39h5270095_0
pip=21.0.1=py39hecd8cb5_0
pixman=0.40.0=hbcb3906_0
prompt-toolkit=3.0.17=pyh06a4308_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-xgboost=1.3.3=py39hecd8cb5_0
pycairo=1.20.0=py39hbe14034_1
pygments=2.8.1=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
python=3.9.4=h88f2d9e_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.9=1_cp39
pytz=2021.1=pyhd3eb1b0_0
pyzmq=20.0.0=py39h23ab428_1
rdkit=2021.03.1=py39h88273a1_0
readline=8.1=h9ed2024_0
reportlab=3.5.67=py39hf37cc50_0
scikit-learn=0.24.1=py39hb2f4e1b_0
scipy=1.6.2=py39hd5f7400_1
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=52.0.0=py39hecd8cb5_0
six=1.15.0=py_0
sqlalchemy=1.4.14=py39h89e85a6_0
sqlite=3.35.4=hce871da_0
termcolor=1.1.0=py_2
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_0
tornado=6.1=py39h9ed2024_0
traitlets=5.0.5=pyhd3eb1b0_0
tzdata=2020f=h52ac0ba_0
wcwidth=0.2.5=py_0
wheel=0.36.2=pyhd3eb1b0_0
xgboost=1.3.3=py39hecd8cb5_0
xlrd=2.0.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.4=h23ab428_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
